Move MatchingOptionSet up to the compiler level
