+++ Sources/_StringProcessing/ConsumerInterface.swift
@@ -38,7 +38,7 @@ extension AST {
   /// A consumer is a Swift closure that matches against
   /// the front of an input range
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction? {
     switch self {
     case .atom(let a):
@@ -72,7 +72,7 @@ extension AST.Atom {
   }
 
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction? {
     // TODO: Wean ourselves off of this type...
     let matchLevel: CharacterClass.MatchLevel = opts.contains(.unicodeScalarSemantics)
@@ -130,7 +130,7 @@ extension AST.Atom {
 
 extension AST.CustomCharacterClass.Member {
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction {
     switch self {
     case .custom(let ccc):
@@ -221,7 +221,7 @@ extension AST.CustomCharacterClass.Member {
 
 extension AST.CustomCharacterClass {
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction {
     // NOTE: Easy way to implement, obviously not performant
     let consumers = try members.map {
@@ -274,7 +274,7 @@ private func consumeScalar(
 
 extension AST.Atom.CharacterProperty {
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction {
     // Handle inversion for us, albeit not efficiently
     func invert(
@@ -344,7 +344,7 @@ extension AST.Atom.CharacterProperty {
 extension Unicode.BinaryProperty {
   // FIXME: Semantic level, vet for precise defs
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction {
     switch self {
 
@@ -508,7 +508,7 @@ extension Unicode.BinaryProperty {
 extension Unicode.POSIXProperty {
   // FIXME: Semantic level, vet for precise defs
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) -> Program<String>.ConsumeFunction {
     // FIXME: semantic levels, modes, etc
     switch self {
@@ -554,7 +554,7 @@ extension Unicode.POSIXProperty {
 extension Unicode.ExtendedGeneralCategory {
   // FIXME: Semantic level
   func generateConsumer(
-    _ opts: AST.MatchingOptionSet
+    _ opts: MatchingOptionSet
   ) throws -> Program<String>.ConsumeFunction {
     switch self {
     case .letter:
